wip: refactoring. dsi factory
